Initial commit
The goal for this app is to be a nice and neat step-by-step repository on how to
set up a modern React Single Page App (SPA). This is more for the sake of my
sanity, as I am extremely forgetful and need some kind of reference log on how
I set up this app. I've also found there a lot of boilerplate apps out there
that have everything all set up and installed for you, but my problem with that
is that I have no idea how it all works, how it's configured and then how to
extend it.

This is my lazy way of building a blog, but not actually building a blog.