Adding Webpack and Yarn
Installed webpack by using yarn. Yarn is an package manager that's very similar
to NPM and is built by Facebook. I like it, because it generally install things
a lot quicker. It also does nice things like generating a lockfile, so whoever
else installs this will be installing the same NPM modules.

This commit also adds Webpack, which is a JS build tool. At the moment, the
current setup and config only just takes our single source file, and copies it
across to the `dist/` directory. We have also updated `index.html` to only look
in `dist/` directory for js files now.

The README has been update with instructions on how install and do this.