Add babel integration to webpack.
The babel-loader added in webpack.config.js means that every file with the .js
or .jsx will be processed with the babel-loader. I am not a 100% sure what
babel-loader is, but as far as I can tell, it's an interface between webpack
to run babel transformation on the specified files.

We have also added a .babelrc file: this tells babel to use the plugins when
it runs. These run in addition to the base babel-core libraries, and will
presumably be invoked when babel-loader is invoked.

We've updated the App.js to use the ES6 arrow functions, so when you run
yarn build again, you'll see in dist/bundle.js that it's actually transformed
the JS code in App.js to use anonymouse functions, so that it's ES5 compatible.