Adding React.
This adds React in the most basic of ways. We are simply just rendering a
simplistic React component, but this shows that all the past commits have tied
together: Babel is importing things properly, and Webpack is traversing the
compile tree to include React, so that when the page loads, all necessary JS
libraries have been loaded, and then it runs our React JS code, to produce
the glorious masterpiece that is our Hello World page.