Adding React components to get imports working.
Here I've just added a bunch of super simplistic components to make sure
that importing with webpack is working all fine. I've had to update the
webpack config to support the `.jsx` extension when it resolves filename
meaning that we can so things like `import House from app/House` and
not have to include src/ or .jsx in that import statement.

There's also another minor update to webpack to also support watching
of source file changes, so that when they do change, webpack will
automatically run and rebuild (you'll still need to hit refresh in the
browser though.