Adding Sass compilation support.
So this is how it works as far as I understand it:
 - Your JS file (App.js in this case) will import a css file, saying
it needs these styles to run/render.
 - Webpack will go to that css file, and traverse the import tree and
get all the CSS files it needs.
 - Webpack will inject a this compiled CSS as a <style> tag when the
compiled JS runs on the page.

This is kind of nice, because it means it will only include and import
the relevant css when it's needed on that page.

The downsides are that it won't be cached, gzipped and will likely have
a flash (or a moment) of unstyled content.

So we have added this ExtractTextPlugin, which will do all of the above,
and then extract out all the css into a single file, `dist/styles.css`
which we then include in our index.html ourselves, but now we can cache
and gzip it all. This is kind of relying on only the master/main app.js
file on importing a single css file, which then goes and imports
all the css files.

I can see some of the benefits of this setup, but it also seems
convoluted. I need to play around with this more to find a more
efficient way to do this.