Serve index.html from dist
This removes our hardcoded references to dist/bundle.js and
dist/styles.css. It now reads a template we define in src/index.html,
and then webpack will inject any css and js links into this file, and
then spits out the resulting html into dist/index.html.

This way, everything we run/serve is coming 100% from the dist directory
now, which will make it easier to manage when it comes to assets.