Adding static assets support (i.e. images)
In the previous commit, we serve everything from the dist/ directory.
Now, this commit adds the file-loader plugin. Anytime webpack encounters
an image or font file when crawling the css or js, it will copy it into
dist/assets as well as appending a hash into the name. It will then
update the js and css files in the dist/ directory to point to
dist/assets/filename.