Add React Router and basic routing examples.
This shows how the "root" component sets up the router, and then child
components can read the URL and links and render the relevant route. The
only messy thing is when we want to send props to a <Route />, we have
to merge it with the props of the calling component to also send through
any route props/params. Not gret, but this realistically wouldn't happen
too often, as each Route should really have it's own info and laod it
via AJAX, or via State and Store.  This same example could also have
been achieved with Route and Route Params and read it from a Store
for any additional info.

This example also shows how to use default State and Props, which is
different from the old getInitialState and getDefaultProps when we
used to do React.createClass. In this case of state, we assign it during
the constructor method. For props, we assign it as a variable _outside_
of the component.

We also use the shiny `${val}` ES6 method, which is like the Ruby string
interpolation method of "Thing: #{var_val}". Finally, we can't loop
through like we used to, and have to create a seperate variable, and
call that inside the main render method.