Adding Express for basic SPA support.
In the previous commit, we added routing, but it was only client side.
When you refreshed, it meant that it hit the server with /app/slytherin
which isn't a supported endpoint in the server, so it would 404 out.
We were also only using a very basic and simple webpack static server.

This commit introduces Express JS as our node server. It takes any
route under /app/ and serves up `dist/index.html`, which renders our
main App page. Then the React Router kicks in, and reads the rest of the
URL parameters, and then renders the relevant view accordingly. This is
a good first step, but to truly make this an isomorphic app, we need to
prerender components on the server side, which I might day later on if
there's time.

To achieve this, we had to ditch webpack-dev-server and instead serve
everything statically from the `dist/` directory via Express. To still
have webpack listening and recompiling for changes though, we've
introduced webpack-dev-middleware and made Express use that. So when
assets changes, Express will tell Webpack, which'll then recompile
assets which'll be available when we refresh.

We've also updated webpack slightly so that it looks for js and css
assets under the `dist/` directory as well as the `/dist/` path in the
URL. This way, we now have some nice and neat namespaces:
 - /app is all client side/react stuff
 - /dist are for assets, such as JS, CSS and Images.

 💥