Clean up routing of components.
Instead of manually hardcoding all 4 routes, we can now just read the
houseName as a parameter from the URL using routes by doing:

`<Route path={`${this.props.match.url}/:houseName`} component={House}/>`

This _doesn't_ pass in any additional props into the component, (the
router will still pass in a bunch) so we can use this this nice and
clean way of creating and initialising routes.

If we did want to pass in our own props in addition to the routing props,
the main `App.jsx` has an example of this. It cleans up slightly the way
we namespace everything under the `/app` path and then initialises the
`HouseContainer` component by mixing in the custom props as well as
any other props available to it.