Add styling when a Tab is selected. Create a Tab component.
This feels a bit hacky, as the current page we are on is only known to
the House component, and the TabBar/Tabs are in the HouseContainer
component. Ideally, there'd be a Store that we can read this information
from. Given the current setup, I think this is okay for now.

The Tab component was created just to clean things up a bit.
If you ever get an error along the lines of
'If you meant to render a collection of children, use an array instead
or wrap the object using createFragment(object) from the React add-ons',
it's because the stateless function that you're looping over has a
collection of params, instead of being bundled in a single JSON object.
Essentially, by doing Tab = ({another, thing}) =>, you're pulling them
out of a JSON object and creating local variables from them. This is
called destructuring arguments in ES6, see:
http://www.saltycrane.com/blog/2016/03/es6-features-used-react-development/#destructuring-function-args