Adding Redux Actions and ActionCreators
Actions are similar to Actions in Flux, but slightly different. An
Action is simply a JSON object of information that tells the state
how it should be modified. In this commit, one example of an action is:

{ type: ENROL_STUDENT, name: name }

Actions always have a type that instruct a Reducer how to modify the
state, as well as any additional information that the Reducer may need
to complete this action. Reducers are the actual function that gets
executed to modify the state (this will be covered in the next commit or
two in greater detail).

However, we need to execute some kind of function to actually send the
Action to the Reducer, which is what we call ActionCreators. Following
on from the above example, we use the following action creator:

export const enrolStudent (name) => {
    let action = { type: ENROL_STUDENT, name: name }
    if (name !== null) { return action; }
}

This is a simple enough function that can do some business logic-y type
things, that when done, returns the action. There's some middleware that
Redux will cover that connects this ActionCreator to the Reducer, which
again, we'll cover in the next commit or two.