Adding Reducers. Adding Redux libraries.
Reducers are our way manipulating the state of an application. They
are given an action, and the _old_ state, and return a _new_ state. They
are "pure" or "stateless" functions: they take in parameters and they
return an output. They do not mutate their parameters, so they never
actually modify the state, but instead, copy it and modify that. The
Redux middleware then takes this new state and updates the global store
(this is covered a bit more in the next commit). In terms of what the
Reducer actually does:

import { ENROL_STUDENT, EXPELL_STUDENT } from './StudentActions'

This is the reason why when we created our Actions
Every Action has a type, and this is the only requirement of an Action.
This is so that in our Reducer, we can read this Type and we know which
Reducer to apply. By exporting the Type names in our actions as
constants, we can safely reference them in our Reducer so if they do
change, things break.

function students(state = [], action) {
    switch (action.type) {
      case ENROL_STUDENT:
        return [...state, action.name]
      case EXPELL_STUDENT:
        return state.filter(n => n !== action.name)
      default:
        return state
    }
}

This is an actual reducer function. Let's take it line by line.

function students(state = [], action) {

This declares our functions and parameters, as well as a default of an
empty array if our state is undefined (aka the first time it runs).

switch (action.type) {

We run different behaviours depending on which action was dispatched.

case ENROL_STUDENT:
  return [...state, action.name]

If the action was to enrol a student, we use the ES6 spread operator
`...` to copy our state, as well as adding in our new student into
this array. This returns a brand new array, without modifying the
state that was passed in, thus keeping the function "pure"

case EXPELL_STUDENT:
  return state.filter(n => n !== action.name)

Similarly, this returns a copy of the array of the state, that excludes
the student with the matching name.

default:
  return state

This is important, as it returns the state if the action doesn't match
any case. So in the case of HIRE_TEACHER and FIRE_TEACHER, it still
returns the state.

const hogwartsApp = combineReducers({
    students,
    teachers
})

This is some Redux magic to combine the various different reducer
functions together so we don't have single giant reducer function,
but many smaller discrete ones.

And that's a reducer! In the next commit, we'll hook it up to the Redux
store and start integrating with our app.