Initialise Store in Root component
The React-Redux library has some nice helper classes for initialisation
and connecting things together.

import { Provider } from 'react-redux'

The Provider we import provides a way for the components, actions and
reducers to talk with each other. It's essentially the equivalent of the
magic that glues Rails together.

import { createStore } from 'redux'
import hogwartsApp from 'app/reducers'
let store = createStore(hogwartsApp)

Initalises the store by taking in the mega reducer we created in the
previous commit.

 ReactDOM.render(
  <Provider store={store}>
    <App />
  </Provider>,
  document.getElementById('main')
);

This is how we initialise the Provider, the store, and how it (is one of
the steps required to) connect the store to the app.