Subscribe HouseContainer to Redux Store
This uses the React-Redux middleware to connect our components and our
store together. It initialises, subscribes and unsubscribes, as well
as setting the props (we don't really use state with Redux).

class HouseContainer extends React.Component {

We no longer export this HouseContainer component. We actually export
a 'connected' version of this component, that is to say, we copy this
copy the component and add the extra bells on whistles, and export that

```
const mapStateToProps = (state) => {
  return { students: state.students }
}
export default connect(mapStateToProps)(HouseContainer)
```

We don't actually use state in components, only props. This is for a
couple of reasons:
 - The main reason is that core principle of Redux is that we have a
 single Store that contains the entire application state. However,
 most components don't need every value in the global store, only a
 small slice. The mapStateToProps function allows us to cut out what
 we need and only subscribe to updates to those values that this
 component cares about.
 - The connect function from react-redux copies the component wraps it,
 so we're not binding to this specific component.  This means we can't
 actually use state, only props, as the copied wrapper component has
 access to state, not the component we're writing.