Connecting Components to Action Creators
This is the final step required to complete our Redux setup. Redux has a
couple of different ways to create Actions using our Action Creators in
our Components. These Actions are then dispatched by Redux to their
relevant Reducers.

export default connect(mapStateToProps, mapDispatchToProps)(HouseCtner)

This magical connect function also binds a function called `dispatch`
into our component. This dispatch function takes a parameter of an
Action, and sends it to the reducer. We have Action Creators to, well,
create Actions, so first we need to import one first:

import { enrolStudent } from 'app/actions/StudentActions';

Import the Action Creator into our component so we can call it. Note
the {} around the enrolStudent, which is the 'destructuring' ES6 method.
Because StudentActions.js exports so many different values, this
destructuring only imports this value, and assigns it into a variable
of the same name.

const addStudent = (event) => {
  if (event.value !== null) {
    return enrolStudent("Ron Weasley")
  }
}

The first way is to define our own method and do things before deciding
to call the Action Creators. This is kind of discouraged, as we should
ideally just call Action Creators and they can then decide if they want
to emit an action or not. However, sometimes you may want to clean up
and parse inputs before sending it to the Action Creator, so this still
okay.

dispatch(enrolStudent("Ron Weasley")),

The dispatch method will:
 - Requires parameters of an Action
 - It dispatches this Action to the Reducer (this is where the Redux
 magic glue comes in)

By executing the enrolStudent() function, which is an Action Creator,
we get an Action returned, which is then immediately dispatched. If we
execute the addStudent() function, it does extra stuff, but ultimately,
still executes enrolStudent(), which returns an Action.

const mapDispatchToProps = (dispatch) => {
  return {
    oldWayToAddStudentClick: (e) => dispatch(addStudent(e))
  }
}
export default connect(mapStateToProps, mapDispatchToProps)(HouseCtner)

We've now added a second parameter to this connect function, which takes
any functions that dispatch to Redux, and maps them into our components
props so that we can reference them. The alternative is to manually
call this.props.dispatch(enrolStudent()) which is perfectly fine, but
it means we have to pass in our props.dispatch in the parameters, which
makes our component really messy, as well as our function. This is just
a neater way to dispatch functions.

And that's it! Redux is now working, end to end, like ~magic!~