Adding basic React snapshot testing
Jest's nicest feature is the React component testing support. It
captures the generated HTML of a rendered React component and saves it
as a snapshot. Subsequent tests then compare the output to this snapshot
so that if the test result is different to the snapshot, it'll fail.
You then just need to compare the result and the snapshot, and determine
if it's an unexpected issue, or what the component should now be looking
like. You can then update your test, or resave the snapshot so that
it's the new expected state.

I've also updated the README.md on how to run tests, and how to debug
them.