Use NPM postinstall to build deployment assets
NPM has a bunch of script hooks that could used in the build lifecycle.
One them is postinstall, which runs immmediately after `npm install` is
run. Because Herokus build scripts run npm install for us, this will
mean that it will run a production version of webpack build for us.

The ugly thing about this is, it means that heroku runs
`npm install --production`. So when it trys to run the webpack build and
any associated tasks, it can't as they're all under devDependencies. To
get around this, I've had to merge dependencies and devDependecies, which
is less than ideal, but I can't quite find a better way to do this, other
than building and committing the dist directory, which is infinitely
worse, as it clutters up and ruins your git history.