Create a production version of the app
The only configuration of running the app was a development version: it
was not using gzipping, minification or uglification. Now, we have two
webpack files, one for dev and one for production, so we can add extra
steps to the build workflow when we do a production build. The server,
`app.js` has been updated to look at the relevant webpack config file,
as well as adding the `compression` plugin, which gzips all requests.

The npm scripts have also been updated so that they're a bit cleaner,
as well as having NODE_ENV variables hardcoded. To run the dev server,
run `yarn dev` and/or `yarn build`. To run the production server,
run `yarn compile`, and then having set NODE_ENV=production in your
environment, run `yarn start`. Or just run
`NODE_ENV=production yarn start`.