Adding Tests for Redux connected components
This commit is a bit of a doozy, but is kind of hard to split apart,
so I apologise in advance.

"moduleDirectories": ["src", "node_modules", "config"],
"moduleNameMapper": {
  "\\.(jpg|jpeg|png|gif....webp)$": "<rootDir>/config/file.mock.js",
  "\\.(css|less)$": "<rootDir>/config/style.mock.js"
}

In `HouseContainer.jsx` file, we import css and image files. We've
configured webpack to handle and process these correctly, but we haven't
for Jest. The above config in the `config/jest.config.js` file will stub
any imports for those kind of asset files to the matching mock file
we've created in the `config/` folder.

"enzyme": "^2.8.2",

Enzyme is a nice helper library for testing React components. It
essentially provides nice utility and wrapper classes for Components,
so you can test the correct method and stuff are called when you click
a button. It also allows for "Full" or "Shallow" rendering.
 - Full rendering does the whole component lifecycle: so things like
`componentDidMount` and stuff are called.
 - Shallow rendering does a lighter version of creating a component,
 which is more suited to unit tests.

<button ref="addStudent" onClick={this.props.preferr.....

Adding a ref so we can target this button in the tests. I think it's
also just a best practice to have refs on user interactable elements.

import renderer from 'react-test-renderer';
import { shallow, mount, render } from 'enzyme';

We need the react-test-renderer so we can do Snapshot testing, and enzyme
so we can do functional tests (click on things, it has a header etc.)

jest.mock('react-router-dom');
jest.mock('app/components/House', () => require('mockComponent'));

Mock out a library or a SubComponent. You can create your own dummy
component that can be a little bit more sophisticated, which is what
the second one does.

In the first HouseContainer.spec.jsx, this is simply a snapshot test,
which renders out the component tree to a file. See commit
 a16fc994096be472e88e1d653652f9382c4f41b1 for more details.

test('It fires an action when clicked', () => {
    let props = {....
    const addStudentClick = jest.fn();
    const wrapper = mount(
      <HouseContainer {...props} preferredWayToAddStudentClick={addStudentClick} />);

   const button = wrapper.ref('addStudent');
   button.simulate('click');
    expect(addStudentClick).toHaveBeenCalled();
});

The second test is doing a bunch of new things. First, we stub out the
`addStudentClick` function, and pass that in as a prop to the component.
We then mount the component with enzyme, so that we can then get the
button, and test it calls our stubbed function when clicked.

Ta-da! Component tested.