Add Redux DevTools via Feature Toggles
Initialises the Redux Store with DevTool support. You'll also need to
have the following Chrome Extension installed:
 - https://chrome.google.com/webstore/detail/redux-devtools/lmhkpmbekcpmknklioeibfkpmmfibljd

Or build and run it from source:
 - https://github.com/zalmoxisus/redux-devtools-extension

There's also now ways to configure and inject variables into the React
App for prod and dev environments at compile time.

new webpack.DefinePlugin({
  PRODUCTION: JSON.stringify(true)
}),

In our webpack config file, we define our variable name and what the
value should be. So when we run `yarn compile`, which uses
`config/webpack.prod.config.js`, it'll read the above. When webpack
compiles, it replaces all instances of `PRODUCTION` with `true` in the
JS, thus feature toggling things on and off in prod or not.

You'll notice it's actually `JSON.stringify(true)` or `"true"` instead
of just `true`. For some reason, webpack is limited to strings when it
does this, and will strip the quotes when it replaces things. So
`"true"` become `true`, and `"'string value'"` becomes 'string value'.
Just something to keep in mind when writing feature toggles!