Adding babel-polyfill, fetch & redux-thunk.
Babel Polyfill adds in support for new ES6/7 browser methods, that not
all browsers support. Babel itself only converts the nice new ES6/7
syntax into ES5 syntax. So, for example, Babel will convert

`thing = p => p.foo()` into `thing = function(p){p.foo()}`

Whereas Babel Polyfill will import a library so that when you call
`fetch()` or `Promise.new()` it'll work in all browsers. We've only
been using normal Babel for now, which is to say we've just been using
the nice new ES6/7 syntax, but in the next commit we're going to be
making API calls with `fetch()`, so we need to make fetch happen.

Fetch is the newer, neater ES6/7 way to make AJAX calls. It returns
Promises and has a nicer API than XMLHttpRequest, and allows for fewer
callbacks to be made.

Finally, redux-thunk is some middleware to allow for Action Creators to
return Promises and functions, instead of just an Action. We'll go into
how and why in the next commit a bit more, but essentially, it allows
us to make API Calls in our Action Creators.