Adding Async/API action creators.
`redux-thunk` is a redux middlexware that allows for asynchronous action
creators. It essentially adds a wrapping Promise function that allows for
dispatches to the reducer, which is useful when you need to do things
like API Calls. Let's breakdown an Async Action Creator:

export const fetchStudents = () => {
  return function (dispatch) {
    dispatch(fetchStudentsBegin());

    return fetch('http://localhost:3000/api/harry_potter')
    .then(response => response.json())
    .then (json => dispatch(fetchStudentsSuccess(json)))
    .catch(json => dispatch(fetchStudentsError(json)));
  }
}

Firstly, you'll notice it returns a function straightaway. This is so
redux-thunk can wrap it in a promise and pass in the `dispatch` method
which talks to the reducer. After this, it makes a fetch request (we'll
cover this in more detail a bit further down). Once successful, or it
errors out, we then dispatch again either a success or an error action.
We could have the 1 action creator that takes in different parameters,
i.e. `fetchStudents("begin")` and `fetchStudents("success")`, but I
think this just boils down to a preference thing.

How do we test this? The expected output of this method is that the
`fetchStudentsBegin` and `fetchStudentsSuccess` methods are dispatched
to our reducer, so that's what we'll do.

import configureMockStore from 'redux-mock-store';
import thunk from 'redux-thunk';
const mockStore = configureMockStore([thunk]);

First, we'll mock out the store (and import a new library to do that).
We need to add thunk to our store so the mock store is setup correctly.

test('it dispatches a successful action', () => {
    const expectedActions = [
      fetchStudentsBegin(),
      fetchStudentsSuccess(data)
    ];
    const store = mockStore({});

    return store.dispatch(fetchStudents())
      .then(() => {
         expect(store.getActions()).toEqual(expectedActions)
       });
 });

Our test case creates an array of actions (by calling the action
creators) that we expect the store to receive so it can send it to the
relevant reducer. Then we "dispatch" the `fetchStudents()` methods,
simulating what's done in our component, and then assert the actions this
method creates, against our expected actions.

The above just covers the async actions creator part, and ignored/removed
the Fetch part. As mentioned in previous commits, Fetch is the new ES6/7
way of making AJAX requests, and replaces XMLHttpRequests (fun fact:
Chrome refactored XMLHttpRequests so that it uses fetch under the hood,
as it's more efficient)

return fetch('http://localhost:3000/api/harry_potter', {
  method: 'GET',
  mode: 'cors',
  headers:{'Access-Control-Allow-Origin':'*'}
}).then(response => {
  let json = response.json();
  if (response.ok) {
    return json;
  } else {
    // Reject the promise if the response is a 500
    return json.then(err => {throw err;});
  }
})
.then (json => dispatch(fetchStudentsSuccess(json)))
.catch(json => dispatch(fetchStudentsError(json)));
}

The first few lines are simple enough, the URL and any options you need
to make the fetch request. The result, or the first `then`, is
called the Response and is a 'stream' object. It makes for nicer
handling/parsing of responses. You can do:
 - `response.json()`, which returns a json object of the body
 - `response.text()`, which returns the body as plaintext (useful for
things like loading html files)
 - `response.blob()` for multimedia things (images, music, etc.)

These all return a Promise Object, so that's why the second `then`
actually dispatches our function to our store. The last one is a `catch`
which is called anytime there's a JS or parsing error. A 500 request
is considered a successful fetch request, as it talked to the server
and came back. That's why in our first Promise callback, we check if
the response is `ok` and if it isn't (if it's a 404 or a 500) then we
throw an error, which rejects the promise and goes to the `catch`
statement.

Testing this is simple enough, `nock` is a fetch mocking library, that
allows for some super clean syntax for mocking a request:

let data = {students: ['Aberforth Dumbledore']};
nock('http://localhost:3000/').get('/api/harry_potter').reply(200, data)
=============================== or ====================================
let data = {code: 500, statusText: 'Oh teh noes! Internal Server Error'};
nock('http://localhost:3000/').get('/api/harry_potter').reply(500, data)

The rest of the changes are just updates to the components so that when
you click the button, it calls this new API Action Creator, as well as
some cleanup of some unused stuff.

This is probably my longest commit message #sorrynotsorry