Create Service Worker
Service workers are JS workers that runs in the background. They can
intercept and cache networks requests, such as AJAX requests, as well as
static assets, allowing your site to run offline.

if ('serviceWorker' in navigator) {
  navigator.serviceWorker.register('/sw.js').then(function(reg) {
    console.log('◕‿◕', reg);
  }, function(err) {
    console.log('ಠ_ಠ', err);
  });
}

To register a service worker for a site, in your normal, front end JS
code, you need to register a Service Worker by passing it a URL to a SW
file. This js file has to be from the same domain, and served over
HTTP and SSL. It'll download this file, and run it in the background
and/or when the page is closed/offline.

When registered under `/dist/sw.js` it can only intercept/cache fetch
requests made under the `/dist/` namespace. Whilst this is useful when
you have multiple namespaces and need multiple service workers, we just
need the one SW. The change to node server allows `/sw.js` to return
`/dist/sw.js`

This `navigator.serviceWorker.register('/sw.js')` can be called everytime
on page load, Chrome has its own versioning smarts to know when to
update the SW, or to download a fresh copy.