Cache assets when the SW installs.
When the service worker installs, we can cache assets on install. Caches
are all promise based.

self.addEventListener('install', (event) => {
  console.log('Installing Service Worker');
  event.waitUntil(
    caches.open(CACHE_NAME)
      .then((cache) => cache.addAll(CACHE_ASSET_REGISTER))
      .then(() => console.log('Successfully installed assets to cache'))
      .catch(() => console.log('Failed to instal assets to cache'))
  );
});

Let's breakdown all the steps involved in install

self.addEventListener('install', (event) => {

The service worker has event listeners that will fire at certain
lifecycle stages. In this case, it'll fire when the browser tries to
install this service workers.

event.waitUntil(

Remember, service workers make extensive usage of promises. The install
event will take in a Promise as a parameter, and wait for that to
complete. If that promise fails, the entire installation of the SW will
fail.

caches.open(CACHE_NAME)
  .then((cache) => cache.addAll(CACHE_ASSET_REGISTER))

We can have multiple caches, and the first line loads up the cache. Then
the next step in the promise chain will add all the assets in our array
into the cache. If any of these assets fail to load/be added to the
cache, then the entire promise chain fails, and thus, the entire SW
installation fails. This is so that we can reliably know which assets
are installed and available for usage in the SW if the page is offline.

self.addEventListener('fetch', (event) => {
  event.respondWith(
    caches.match(event.request)
          .then((response) => response || fetch(event.request))
  );
});

This final piece is a listener for any `fetch` events, aka any network
requests the browser makes. All network requests by chrome now use fetch
including XMLHttpRequest and any requests for assets and stuff. This
method simply checks if there's any request that matches something
in the cache, and if there is, return that response. Otherwise,
commit an actual fetch and make the proper network request.

Now, our SW is installed and ready for usage!

If you load the page, and then kill your server, and refresh, the page
should still load, as the SW is now caching assets and returning those.