Fix specs by creating a common constants file
`nock` does not support mocking of same origin fetch requests, as it is
meant to be more mocking fetch on node, not server side. That is to say,
you can't do `nock('/').get('/api/request/')`. So I've pulled out the
the URL path into a utility file, and then mocked the tests so the
endpoint is now `http://localhost:3000` instead, which is a valid thing
for nock to mock (heh).