Adding a manifest.json for better PWA support
A manifest.json is a config file for Chrome to read to better make the
site a PWA on mobile devices. It contains information like a name,
icon, colour, starting url so that when you save the site to the home
screen, it can 'simulate' looking like a normal native app.

Because we've added a manifest and an icon (well, icons, for differing
pixel densities) into our config folder, we need to copy this across
to the `dist/` folder so that they can be statically served from there.

The final piece is in `src/app/index.html` which specifies our manifest
path via the following attribute in the page head:

<link rel="manifest" href="/dist/manifest.json">